Expose MockAggregator in public module, testing
